requirements.txt
requests
tqdm
GitPython